% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.group.R
\name{analyses.group}
\alias{analyses.group}
\title{Analyses per group}
\usage{
analyses.group(axes, base = 1, groups, ...)
}
\arguments{
\item{axes}{a list of vector coordinates (typically from \code{get.axes})}

\item{base}{which element in \code{"beer"} to use as the base vector (by default the first element is used).}

\item{groups}{which element in \code{"beer"} to use analyse (if left missing, all but the base are analysed).}

\item{...}{any arguments to be passed to \code{\link{vector.diff}}.}
}
\description{
Wrapper for \code{\link{vector.diff}} to run the analyses per groups
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
