% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.covar.R
\name{as.covar}
\alias{as.covar}
\title{Use covar}
\usage{
as.covar(metric, ...)
}
\arguments{
\item{metric}{The metric to adapt as a covar one.}
}
\description{
Adapts a dispRity.metric to use a covar element from a dispRity object
}
\examples{
data(BeckLee_mat50)
## Using the centroids metric on a normal disparity object
dispRity(BeckLee_mat50, metric = centroids)
## Same for the mean centroids
dispRity(BeckLee_mat50, metric = c(mean, centroids))

## Using it on a dispRity object containing a covar part
dispRity(BeckLee_mat50, metric = as.covar(centroids))
## Same for the mean centroids
dispRity(BeckLee_mat50, metric = c(mean, as.covar(centroids)))

}
\seealso{

}
\author{
Thomas Guillerme
}
