% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beer-package.R
\docType{package}
\name{beer-package}
\alias{beer}
\alias{beer-package}
\title{beer: an R package for quantifying antibody reactivities for PhIP-Seq data}
\description{
The core functions of beer are \code{\link{brew()}} and
\code{\link{\runEdgeR()}}. To learn more about \code{beer}, see the vignette
using \code{browseVignettes(package = "beer")}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/athchen/beer/}
  \item Report bugs at \url{https://github.com/athchen/beer/issues}
}

}
\author{
\strong{Maintainer}: Athena Chen \email{achen70@jhu.edu} (\href{https://orcid.org/0000-0001-6900-2264}{ORCID})

Authors:
\itemize{
  \item Rob Scharpf \email{rscharpf@jhu.edu}
  \item Ingo Ruczinski \email{ingo@jhu.edu}
}

}
\keyword{internal}
