% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_data_response.R
\name{check_for_data_response}
\alias{check_for_data_response}
\title{Check if the response from the API call includes data.}
\usage{
check_for_data_response(api_response)
}
\arguments{
\item{api_response}{}
}
\value{
A logical object indicating if data is part of the response object
}
\description{
Check if the response from the API call includes data.
}
\keyword{internal}
