% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.mini.chains.R
\name{combine.mini.chains}
\alias{combine.mini.chains}
\title{combine.mini.chains}
\usage{
combine.mini.chains(mini.chains)
}
\arguments{
\item{mini.chains}{a "mini.chains" objects}
}
\description{
Combine the mini chains results into a big MCMCglmm object (MCMCglmm)
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
