% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar.projections.wrapper.R
\name{covar.projections.wrapper}
\alias{covar.projections.wrapper}
\title{Covar projection analyses wrapper}
\usage{
covar.projections.wrapper(
  data,
  type,
  base,
  average,
  major.axis = 1,
  level = 0.95,
  measure = c("position", "distance", "degree")
)
}
\arguments{
\item{data}{a \code{dispRity} object containing a covar component}

\item{type}{either \code{"within"} for the projection of the group onto itself or \code{"between"} for the projections of the groups onto the \code{base} (see details)}

\item{base}{optional, if \code{type = "between"}, a \code{character} string for the name of the subset in data to use as a baseline (e.g. the phylogeny) or if \code{type = "within"} the major axis to use as a base (see details).}

\item{average}{optional, if \code{type = "within"}, a the function for averaging the major axes (see details)}

\item{major.axis}{which major axis to use (default is \code{1})}

\item{level}{the confidence interval to estimate the major axis (default is \code{0.95})}

\item{measure}{which parameters to measure from the projection. Can be any of the following \code{"c(position, distance, degree)"} (default) for respectively the distance on the projection, the distance from the projection and the angle of the projection.}
}
\description{
Wrapper function for a covar projection analyses
}
\details{
Effectively, the wrapper runs either of the following function (simplified here):
\itemize{
\item if \code{type = "between"}: \code{dispRity(data, metric = projections.covar, between.groups = TRUE)} for the pairwise projections between each subset in \code{data}.
\item if \code{type = "within"}: \code{dispRity(data, metric = projections, point1 = axis[1,], point2 = axis[1,])} for the projections between each subset in \code{data} onto an axis.
}
If \code{base} is specified:
\itemize{
\item \code{type = "between"} will run pairs between each subset and \code{base} (instead of the full pairwise analyses).
\item \code{type = "within"} will run the projection of each subset onto the major axis from \code{base} rather than its own.
}
}
\examples{

}
\seealso{
\code{\link[dispRity]{projections}}  \code{\link[dispRity]{axis.covar}}  \code{\link[dispRity]{dispRity}}
}
\author{
Thomas Guillerme
}
