% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.mini.chains.R
\name{diagnose.mini.chains}
\alias{diagnose.mini.chains}
\title{diagnose.mini.chains}
\usage{
diagnose.mini.chains(mini.chains)
}
\arguments{
\item{mini.chains}{a MCMCglmm result}
}
\description{
Diagnose a set of mini chains results
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
