% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRity.covar.R
\name{dispRity.covar}
\alias{dispRity.covar}
\title{Temporary wrapper for dispRity}
\usage{
dispRity.covar(data, metric, between.groups = FALSE, ...)
}
\arguments{
\item{data}{the data with covar}

\item{metric}{the metric}

\item{between.groups}{(for group analyses or tip analyses)}
}
\description{
TO BE PROPERLY INCLUDED IN dispRity
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
