% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkCounts}
\alias{.checkCounts}
\title{Function to check that the counts matrix only contains integers}
\usage{
.checkCounts(object)
}
\arguments{
\item{object}{PhIPData object}
}
\value{
nothing if all counts are integers, and error otherwise
}
\description{
Function to check that the counts matrix only contains integers
}
