% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkOverwrite}
\alias{.checkOverwrite}
\title{Function to check whether an assay will be overwritten}
\usage{
.checkOverwrite(object, assay.names)
}
\arguments{
\item{object}{PhIPData object}

\item{assay.names}{character vector of assay names}
}
\value{
logical vector indicating whether data in an assay will be
overwritten
}
\description{
If the an assay is not specified (e.g. with \code{NA}), then
\code{.checkOverwrite()} will return \code{FALSE} (rather than NA).
}
