% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR.R
\name{.edgeR_one}
\alias{.edgeR_one}
\title{Run edgeR for one sample against all the beads-only samples.}
\usage{
.edgeR_one(object, sample, beads, common.disp, tagwise.disp, trended.disp)
}
\arguments{
\item{object}{\code{\link[PhIPData]{PhIPData}} object}

\item{sample}{sample name of the sample to compare against beads-only samples}

\item{beads}{sample names for beads-only samples}

\item{common.disp}{edgeR estimated common disperion parameter}

\item{trended.disp}{edgeR estimated trended dispersion parameter}

\item{tagwise.dsip}{edgeR estimated tagwise dispersion parameter}
}
\value{
list with sample name, log2 fc estimate, and log10 p-value
}
\description{
Run edgeR for one sample against all the beads-only samples.
}
