% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAB.R
\name{.getAB_MOM_prop}
\alias{.getAB_MOM_prop}
\title{Helper function to derive MOM estimates of a, b from a vector of proportions}
\usage{
.getAB_MOM_prop(
  prop,
  offsets = c(mean = 1e-08, var = 1e-08),
  lower = 1,
  upper = Inf,
  ...
)
}
\arguments{
\item{prop}{vector of proportions.}

\item{offsets}{vector defining the offset to use when the mean and/or
variance are zero.}

\item{lower}{lowerbound for the shape parameters.}

\item{upper}{upper bound for the shape parameters.}

\item{...}{parameters passed to \code{[base::mean]} and
\code{[stats::var]}.}
}
\value{
a data frame with MOM estimates of a, b
}
\description{
Helper function to derive MOM estimates of a, b from a vector of proportions
}
