% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAB.R
\name{.getABMLEProp}
\alias{.getABMLEProp}
\title{Helper function to derive MLE estimates of a, b from a vector of proportions}
\usage{
.getABMLEProp(
  prop,
  prop.offset = 1e-08,
  optim.method = "default",
  lower = 1,
  upper = Inf
)
}
\arguments{
\item{prop}{vector of proportions.}

\item{prop.offset}{offset to use when the proportion of reads is 0.}

\item{optim.method}{optimization method passed to \code{[stats::optim]}.}

\item{lower}{lower bound for the shape parameters.}

\item{upper}{upper bound for the shape parameters.}
}
\value{
a data frame of MLE estimates of a, b
}
\description{
Helper function to derive MLE estimates of a, b from a vector of proportions
}
\seealso{
\code{[stats::optim]} for available optimization methods
}
