% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAB.R
\name{.getABMOM}
\alias{.getABMOM}
\title{Wrapper function to derive MOM estimates of a, b from beads-only samples}
\usage{
.getABMOM(
  object,
  offsets = c(mean = 1e-08, var = 1e-08),
  lower = 1,
  upper = Inf,
  ...
)
}
\arguments{
\item{object}{a \code{\link[PhIPData]{PhIPData}} object.}

\item{offsets}{vector defining the offset to use when the mean and/or
variance are zero.}

\item{lower}{lower bound for the shape parameters.}

\item{upper}{upper bound for the shape parameters.}

\item{...}{parameters passed to \code{[base::mean]} and
\code{[stats::var]}.}
}
\value{
a data frame with MOM estimates of a, b
}
\description{
Wrapper function to derive MOM estimates of a, b from beads-only samples
}
