% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew.R
\name{.tidyAssayNames}
\alias{.tidyAssayNames}
\title{Clean-up specified assay names}
\usage{
.tidyAssayNames(assay.names)
}
\arguments{
\item{assay.names}{named list specifying where to store each assay.}
}
\value{
tidied list of assay.names
}
\description{
Tidy inputs related to `assay.names`. Supplies default values for
missing parameters and ensures that all required parameters are present.
}
