% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew.R
\name{.tidyInputsPrior}
\alias{.tidyInputsPrior}
\title{Clean up inputs for prior estimation}
\usage{
.tidyInputsPrior(prior.params, object, beads.args)
}
\arguments{
\item{prior.params}{named list of prior parameters}

\item{object}{PhIPData object}

\item{beads.args}{parameters used to estimate a_0, b_0}
}
\value{
tidied list of prior parameters.
}
\description{
Tidy inputs related to `prior.parameters`. Supplies default values for
missing parameters and ensures that all required parameters are present.
}
