% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew.R
\name{.tidyInputsSE}
\alias{.tidyInputsSE}
\title{Clean up inputs for identifying super-enriched peptides}
\usage{
.tidyInputsSE(se.params, beads.prior)
}
\arguments{
\item{se.params}{named list of parameters for super-enriched estimation}

\item{beads.prior}{data.frame with beads-only parameters}
}
\value{
tidied list of parameters for identifying super-enriched peptides.
}
\description{
Tidy inputs related to `se.params`. Supplies default values for
missing parameters and ensures that all required parameters are present.
}
