% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.tidyParallel}
\alias{.tidyParallel}
\title{Function to tidy inputs pertaining to parallelization}
\usage{
.tidyParallel(parallel)
}
\arguments{
\item{parallel}{list of parameters passed to the
\code{[future::plan]} function.}
}
\value{
tidied list of parameters passed to \code{[future::plan]}
}
\description{
This function ensures that the specified plan is in the list of valid
plans for \code{[future::plan]}. The function also ensures that
the maximum number of workers is one less that number of available number
of cpus.
}
