% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.replicate.R
\name{eval.verbose}
\alias{eval.verbose}
\title{Optimised replicates}
\usage{
eval.verbose(fun, msg = ".")
}
\arguments{
\item{function}{The function to execute}

\item{minimum}{The minimum of replicates to run (default is 20)}

\item{maximum}{Optional, a maximum of iterations (default is Inf).}

\item{diagnose}{The diagnosis function}

\item{stop.variance}{The percentage of variance change for stopping (default is 0.05).}

\item{increment}{The percentage of increment (if left empty, the value from stop.variance is used).}

\item{verbose}{Whether to be verbose (\code{TRUE}) or not (\code{FALSE}; default)}

\item{parallel}{Whether to use the \code{\link[future.apply]{future_replicate}} function for parallelisation (\code{TRUE}) or not (\code{FALSE}; default).}

\item{par.plan}{If \code{parallel = TRUE}, which plan to use (see \code{\link[future]{plan}}; by default, the plan is "future::multisession").}

\item{...}{Any optional arguments to be passed to the function.}
}
\description{
A wrapper for replicating a based on results variance
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
