% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_beer.R
\name{find_beer}
\alias{find_beer}
\title{Find a beer in BreweryDB}
\usage{
find_beer(api_key, name = NULL, ids = NULL, style_id = NULL,
  organic = NULL, status = "verified", data_frame = TRUE)
}
\arguments{
\item{api_key}{a charater string containing a BreweryDB API key.
See \url{http://www.brewerydb.com/developers} to obtain an API key.}

\item{name}{a character string containing the name of a beer.}

\item{ids}{a character string or vector of character strings containing
beer ids.}

\item{style_id}{a character string providing BreweryDB's ID for a beer style.}

\item{organic}{a character string, either \code{'Y'} or \code{'N'} indicating
if the beer is organic.}

\item{status}{a character string indicating the status of an entity in
BreweryDB.  Defaults to \emph{verified}.  Other possible status are
\emph{new_new_unverified}, \emph{update_pending}, \emph{delete_pending} and
\emph{deleted}.}

\item{data_frame}{a logical argument indicating if the function should return
a data frame.  Defaults to \code{TRUE}.}
}
\value{
By default the function returns a data frame.  If
\code{data_frame = FALSE} the function returns a list of responses - 1 for
each page of results - created by
\code{\link[httr]{GET}}
}
\description{
Find a beer in BreweryDB
}
\examples{

}
