% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_beer_ingredients.R
\name{find_beer_ingredients}
\alias{find_beer_ingredients}
\title{Find ingridents used to make a beer.}
\usage{
find_beer_ingredients(api_key, id, data_frame = TRUE)
}
\arguments{
\item{api_key}{a charater string containing a BreweryDB API key.
See \url{http://www.brewerydb.com/developers} to obtain an API key.}

\item{data_frame}{a logical argument indicating if the function should return
a data frame.  Defaults to \code{TRUE}.}

\item{ida}{character string containing a beer id.}
}
\value{
By default the function returns a data frame.  If
\code{data_frame = FALSE} the function returns a list of responses - 1
for each page of results - created by \code{\link[httr]{GET}}
}
\description{
Find ingridents used to make a beer.
}
\examples{
# Do Not Run
# find_beer_ingredients(key, id = 'WHQisc')
}
