% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_brewery.R
\name{find_brewery}
\alias{find_brewery}
\title{Find a brewery in BreweryDB}
\usage{
find_brewery(api_key, name = NULL, ids = NULL, established = NULL,
  status = "verified", data_frame = TRUE)
}
\arguments{
\item{api_key}{a charater string containing a BreweryDB API key.
See \url{http://www.brewerydb.com/developers} to obtain an API key.}

\item{name}{a character string containing the name of a brewery.}

\item{ids}{a character string or vector of character strings containing
brewery ids.}

\item{established}{a character string containing the year indicating the year
a brewery or breweries where established.  Must follow the format 'YYYY'.}

\item{status}{a character string indicating the status of an entity in BreweryDb.  Defaults to \emph{verified}.  Other possible status are
\emph{new_new_unverified}, \emph{update_pending}, \emph{delete_pending} and \emph{deleted}}

\item{data_frame}{a logical argument indicating if the function should return
a data frame.  Defaults to \code{TRUE}.}
}
\value{
By default the function returns a data frame.  If
\code{data_frame = FALSE} the function returns the response object created
by \code{\link[httr]{GET}}
}
\description{
Find a brewery in BreweryDB
}
