% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_catagories.R
\name{find_categories}
\alias{find_categories}
\title{Find Beer Categories.}
\usage{
find_categories(api_key, id = NULL, name = NULL, data_frame = TRUE)
}
\arguments{
\item{api_key}{a charater string containing a BreweryDB API key.
See \url{http://www.brewerydb.com/developers} to obtain an API key.}

\item{id}{a character string providing the BreweryDB ID of a specific beer
category.}

\item{name}{a charater string providing the name of a specific beer category.}

\item{data_frame}{a logical argument indicating if the function should return
a data frame.  Defaults to \code{TRUE}}
}
\value{
By default the function returns a data frame.  If
\code{data_frame = FALSE} the function returns the response object created
by \code{\link[httr]{GET}}
}
\description{
Find Beer Categories.
}
\examples{

}
