% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_hops.R
\name{find_hops}
\alias{find_hops}
\title{Find all hop styles in Brewery DB}
\usage{
find_hops(api_key, data_frame = TRUE)
}
\arguments{
\item{api_key}{a charater string containing a BreweryDB API key.
See \url{http://www.brewerydb.com/developers} to obtain an API key.}

\item{data_frame}{a logical argument indicating if the function should return
a data frame.  Defaults to \code{TRUE}.}
}
\value{
By default the function returns a data frame.  If
\code{data_frame = FALSE} the function returns a list of responses - 1
for each page of results - created by \code{\link[httr]{GET}}
}
\description{
Find all hop styles in Brewery DB
}
\examples{
# Do Not Run
hop_df <- find_hops(api_key)
hop_list <- find_hops(api_key, data_frame = FALSE)

}
