% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_locations.R
\name{find_locations}
\alias{find_locations}
\title{Find Breweries By Location}
\usage{
find_locations(api_key, locality = NULL, region = NULL,
  postal_code = NULL, status = "verified", data_frame = TRUE)
}
\arguments{
\item{api_key}{a charater string containing a BreweryDB API key.
See \url{http://www.brewerydb.com/developers} to obtain an API key.}

\item{locality}{a character string indicating the location of the breweries.
In the US this corespods to a city.}

\item{region}{a character string indicating the location of the breweries.
In the US this corespods to a state}

\item{postal_code}{a character string providing the postal code of a location.}

\item{status}{a character string indicating the status of an entity in
BreweryDb.  Defaults to \emph{verified}.  Other possible status are
\emph{new_new_unverified}, \emph{update_pending}, \emph{delete_pending}
and \emph{deleted}}

\item{data_frame}{a logical argument indicating if the function should return
a data frame.  Defaults to \code{TRUE}.}
}
\value{
By default the function returns a data frame.  If
\code{data_frame = FALSE} the function returns the response object created
by \code{\link[httr]{GET}}
}
\description{
Find Breweries By Location
}
