% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_styles.R
\name{find_styles}
\alias{find_styles}
\title{Find Beer Styles.}
\usage{
find_styles(api_key, style_id = NULL, style_name = NULL,
  category_id = NULL, category_name = NULL, data_frame = TRUE)
}
\arguments{
\item{api_key}{a charater string containing a BreweryDB API key.
See \url{http://www.brewerydb.com/developers} to obtain an API key.}

\item{style_id}{a character string providing the BreweryDB ID of a specific
beer style.}

\item{style_name}{a character string providing the name of specific beer
style.}

\item{category_id}{a character string providing the BreweryDB ID of a
specific beer category.}

\item{category_name}{a charater string providing the name of a specific beer
category.}

\item{data_frame}{a logical argument indicating if the function should return
a data frame.  Defaults to \code{TRUE}.}
}
\value{
By default the function returns a data frame.  If
\code{data_frame = FALSE} the function returns a list of responses - 1 for each page of results - created by
\code{\link[httr]{GET}}
}
\description{
Find Beer Styles.
}
\examples{

}
