% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.axes.R
\name{get.axes}
\alias{get.axes}
\title{get.axes}
\usage{
get.axes(beer, axes = 1)
}
\arguments{
\item{beer}{the \code{"beer"} object}

\item{axes}{which major axes to get (default is \code{1})}
}
\description{
Get the major axes from a list of covar matrices
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
