% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.covar.R
\name{get.covar}
\alias{get.covar}
\title{Get covariance matrices}
\usage{
get.covar(data, n, simplify = TRUE)
}
\arguments{
\item{data}{The MCMCglmm object}

\item{n}{The number of samples}

\item{simplify}{logical, whether to output the results as a matrix of lists (simplify = FALSE) where rows are the levels and columns are the replicates or whether to return it as a list containing n levels elements each containing the replicates (simplify = TRUE).}
}
\description{
Get a sub-sample of the covariance matrices from a MCMCglmm object and the associated MME posterior (Solution) sorted by levels
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme, Gavin Thomas
}
