% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_pages.R
\name{get_multiple_pages}
\alias{get_multiple_pages}
\title{Process multi page responses from an API call.}
\usage{
get_multiple_pages(number_of_pages, api_response)
}
\arguments{
\item{number_of_pages}{an integer.  The total number of pages returned by the
current API call.}

\item{api_response}{a response object created by \code{\link[httr]{GET}}.
More specifically it is the respones from the original API call which
returned multiple result pages.}
}
\value{
returns a named list with a length equal to the number of pages in the
response.  Each item in the list is a response object created
by \code{\link[httr]{GET}}.
}
\description{
Process multi page responses from an API call.
}
\keyword{internal}
