% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.results.R
\name{group.results}
\alias{group.results}
\title{group.results}
\usage{
group.results(beer, what)
}
\arguments{
\item{beer}{the \code{"beer"} object}

\item{what}{which results to get}
}
\description{
Get results per groups
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
