% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_parse.R
\name{json_parse}
\alias{json_parse}
\title{Parse JSON response and return a data frame.}
\usage{
json_parse(api_content)
}
\arguments{
\item{api_content}{an object of class "response" from \code{\link[httr]{GET}}}
}
\value{
returns a data frame
}
\description{
Parse JSON response and return a data frame.
}
\keyword{internal}
