% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.mini.chains.R
\name{make.mini.chains}
\alias{make.mini.chains}
\title{make.mini.chains}
\usage{
make.mini.chains(data, dimensions, trees, model, parameters, priors)
}
\arguments{
\item{data}{a full ordinated matrix}

\item{dimensions}{the list of dimensions to include}

\item{trees}{the list of trees to run the models on}

\item{model}{the model}

\item{parameters}{the tunning parameters for the MCMCglmm}

\item{priors}{the priors for the MCMCglmm}
}
\description{
Making a list of mini.chains object (MCMCglmm primers)
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
