\name{multiplot}
\alias{multiplot}
\title{Plots multiple}
\usage{
  multiplot(plots, cols = 1, plot.title = NULL)
}
\arguments{
  \item{plots}{Passes all plots as a list}

  \item{cols}{Number of columns to be used}

  \item{plot.title}{Name of the entire plot}
}
\description{
  its awesome, real nice
}

