% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.replicate.R
\name{optim.replicate}
\alias{optim.replicate}
\title{Optimised replicates}
\usage{
optim.replicate(
  input.fun,
  diagnose,
  summarise,
  minimum = 20,
  maximum = Inf,
  stop.variance = 0.05,
  increment,
  verbose = FALSE,
  parallel = FALSE,
  par.plan = "future::multisession",
  bkp.path,
  bkp.name,
  ...
)
}
\arguments{
\item{input.fun}{The function to execute. The function should output a single value or a vector or a list of values.}

\item{diagnose}{The diagnosis function should intake a vector of values.}

\item{summarise}{Optional, if the output of \code{input.fun()} is complex, a function to summarise that output and pass it to \code{diagnose()}.}

\item{minimum}{The minimum of replicates to run (default is 20)}

\item{maximum}{Optional, a maximum of iterations (default is Inf).}

\item{stop.variance}{The percentage of variance change for stopping (default is 0.05).}

\item{increment}{The percentage of increment (if left empty, the value from stop.variance is used).}

\item{verbose}{Whether to be verbose (\code{TRUE}) or not (\code{FALSE}; default)}

\item{parallel}{Whether to use the \code{\link[future.apply]{future_replicate}} function for parallelisation (\code{TRUE}) or not (\code{FALSE}; default).}

\item{par.plan}{If \code{parallel = TRUE}, which plan to use (see \code{\link[future]{plan}}; by default, the plan is "future::multisession").}

\item{bkp.path}{Optional, a \code{"character"} string to be the path to a backup file. See details.}

\item{bkp.name}{Optional, if \code{bkp.path} is not missing, a \code{"character"} for the name of the file to save (or to load). See details.}

\item{...}{Any optional arguments to be passed to the function.}
}
\value{
A list of:
\itemize{
\item{results}: a \code{matrix} with the summarised output from \code{fun()} for every set of iterations;
\item{diagnoses}: a \code{matrix} with changes in variances after each diagnosis;
\item{outputs}: a \code{list} of the outputs of each iteration of \code{fun()}.
}
}
\description{
A wrapper for replicating a based on results variance
}
\details{
When \code{bkp} is specified, the algorithm first looks if a file with \code{bkp.name} exists (by default, the name is \code{"optim.replicate.bkp"}). If not, the results will be saved after each series of replicates into this file. If it exist, the initial steps are skipped and the algorithm continues from the last backup point. \emph{NOTE} that this option slows down the simulations.
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
