% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.analyses.group.R
\name{plot.analyses.group}
\alias{plot.analyses.group}
\title{Plot analyses per group}
\usage{
\method{plot}{analyses.group}(results, what, col, ...)
}
\arguments{
\item{results}{the results of an analyses per group (typically from \code{\link{analyses.group}}).}

\item{what}{which results to plot. By default everything is plotted (e.g. \code{c("angle", "projection", "rejection")}).}

\item{col}{a colour vector (for the levels).}

\item{...}{any plotting options to be passed to \code{\link{graphics}{boxplot}}}
}
\description{
Plotting the results of a per group analyses
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
