% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.analyses.tip.R
\name{plot.analyses.tip}
\alias{plot.analyses.tip}
\title{Plot analyses tips}
\usage{
\method{plot}{analyses.tip}(results, what, col, group, ...)
}
\arguments{
\item{results}{the results of an analyses per group (typically from \code{\link{analyses.group}}).}

\item{what}{which results to plot. By default everything is plotted (e.g. \code{c("projections", "rejections")}).}

\item{col}{a colour vector (for the levels).}

\item{group}{optional, either a list factors for grouping the data or a phylogenetic tree.}

\item{...}{any plotting options to be passed to \code{\link{graphics}{boxplot}}}
}
\description{
Plotting the results of the analyses per tips
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
