% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.axes.R
\name{plot.axes}
\alias{plot.axes}
\title{plot.axes}
\usage{
\method{plot}{axes}(
  data,
  dimensions = c(1, 2),
  centre = "intercept",
  col,
  add = TRUE,
  transparent.scale = 1,
  ...
)
}
\arguments{
\item{data}{A list of axes coordinates (typically from \code{\link{get.axes}})}

\item{dimensions}{which dimensions to plot (default is \code{c(1,2)}).}

\item{centre}{optional, the centre of the ellipses (see details)}

\item{col}{a colour vector.}

\item{add}{logical, whether to add the lines to an existing plot (\code{TRUE}, default) or not (\code{FALSE}).}

\item{transparent.scale}{A numerical value to fade the colours (1 = no fade, 0 = no colours).}

\item{...}{any plotting options to be passed to \code{\link{graphics}{lines}}}

\item{axis}{which major axes to plot (default is \code{1}).}

\item{scale.axes}{logical, whether to set both axes on the same scale (\code{TRUE}) or not (\code{FALSE}, default).}
}
\description{
Plotting axes from an ellipse
}
\details{
\code{centre} can be one of the following:
\itemize{
\item \code{"intercept"} to use the estimated intercept is used (i.e. \code{MCMCglmm$Sol});
\item a \code{numeric} of one or 2 elements to be used as the x/y coordinates of the centre
\item a \code{function} to estimate the central tendency of the estimated intercepts (e.g. \code{mean})
}
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
