% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dots.R
\name{plot.dots}
\alias{plot.dots}
\title{Internal: plot.dots}
\usage{
\method{plot}{dots}(data, dimensions, main_default, dimensions.prefix = "Dim.", ...)
}
\arguments{
\item{data}{for x/y limits}

\item{dimensions}{for x/y labels}

\item{main_default}{for a main}

\item{dimensions.prefix}{for a prefix on the labels}
}
\description{
And ... arguments for plot functions
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
