% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.groups.R
\name{plot.groups}
\alias{plot.groups}
\title{plot.groups}
\usage{
\method{plot}{groups}(beer, dimensions, ..., add)
}
\arguments{
\item{beer}{the \code{"beer"} object.}

\item{...}{any plotting options to be passed to \code{\link{graphics}{lines}}}

\item{add}{logical, whether to add the lines to an existing plot (\code{TRUE}, default) or not (\code{FALSE}).}
}
\description{
Plotting axes from an ellipse
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
