% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.space.R
\name{plot.space}
\alias{plot.space}
\title{plot.space}
\usage{
\method{plot}{space}(beer, dimensions, ..., add)
}
\arguments{
\item{beer}{the \code{"beer"} object.}

\item{dimensions}{which dimensions to plot (default is \code{c(1,2)}).}

\item{...}{any plotting options to be passed to \code{\link{graphics}{lines}}}

\item{add}{logical, whether to add the lines to an existing plot (\code{TRUE}, default) or not (\code{FALSE}).}
}
\description{
Plotting axes from an ellipse
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
