% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.covar.R
\name{projections.covar}
\alias{projections.covar}
\title{level 1 between group metric for dispRity (projections wrapper for covar - MCMCglmm objects)}
\usage{
projections.covar(
  matrix,
  matrix2,
  measure = "position",
  scaled = TRUE,
  axis = 1,
  level = level
)
}
\description{

}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
