% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.mini.chains.R
\name{run.mini.chains}
\alias{run.mini.chains}
\title{run.mini.chains}
\usage{
run.mini.chains(mini.chains, parallel, path)
}
\arguments{
\item{mini.chains}{a "mini.chains" objects}

\item{path}{}
}
\description{
Runs a list of mini.chains objects (MCMCglmm)
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
