% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sauron.plot.R
\name{sauron.plot}
\alias{sauron.plot}
\title{sauron.plot}
\usage{
sauron.plot(
  data,
  n,
  points = TRUE,
  major.axes = NULL,
  ellipses = NULL,
  level = 0.95,
  dimensions = c(1, 2),
  centres = colMeans,
  transparent.scale = 0,
  legend = FALSE,
  legend.pos = "topleft",
  ...
)
}
\arguments{
\item{data}{an MCMCglmm dispRity object}

\item{n}{only selecting a random subset of the posteriors (if left missing, all the posteriors are used)}

\item{points}{logical, whether to plot the points (\code{TRUE}; default) or not (\code{FALSE}).}

\item{major.axes}{can be \code{NULL} or \code{FALSE} for no axes, \code{"all"} or \code{TRUE} for displaying the \code{n} posteriors or a \code{function} for summarising the major axes (e.g. \code{major.axes = mean}.}

\item{ellipses}{can be \code{NULL} or \code{FALSE} for no ellipses, \code{"all"} or \code{TRUE} for displaying the \code{n} posteriors or a \code{function} for summarising the ellipses (e.g. \code{ellipses = mean}.}

\item{level}{the confidence interval level of the major axes and ellipses (default is \code{0.95});}

\item{dimensions}{which dimensions (default is 1:2), could be 3D.}

\item{centres}{optional, a way to determine ellipses or major axes centres. Can be either a function (default is \code{colMeans}), a vector or a list of coordinates vectors or "intercept". See details.}

\item{transparent.scale}{if multiple major axes/ellipses are plotted, a scaling factor for the transparency. If left empty, the transparency is set to \code{1/n} or \code{0.1} (whichever is higher).}

\item{legend}{whether to add the automatic legend (\code{TRUE}) or not (\code{FALSE}; default).}

\item{legend.pos}{the coordinates of where to add the legend (if \code{legend = TRUE}). Can a numeric vector or a character string like \code{"topleft"} (default), \code{"bottomright"}, etc..}

\item{...}{graphic parameters}
}
\description{
plot a MCMCglmm.disparity object
}
\details{
The argument \code{centres} allows to determine how to calculate the centre of each ellipses or major axes. The argument can be either:
\itemize{
\item A function to calculate the centre from a group like the default \code{colMeans} function that calculates the centroid coordinates of each group;
\item A numeric value to be replicated as the coordinates for the centre of each group (e.g. \code{centres = 0} sets all the centres at the coordinates \code{c(0,0,0,...)}); or a vector of numeric values to be directly used as the coordinates for each group (e.g. \code{centres = c(1,2,3)} sets all the centres at the coordinates \code{c(1,2,3)}); or a list of numeric values or numeric vectors to be used as the coordinates for the centres of each group;
\item code{"intercept"} for using the estimated posterior intercept for each sample.
}

\emph{NOTE} that if the input contains more dimensions than the visualised dimensions (by default \code{dimensions = c(1,2)}) the ellipses and major axes are projected from an n-dimensional space onto a 2D space which might make them look "off".
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
