% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{styles}
\alias{styles}
\title{Styles}
\format{a data.frame}
\source{
\url{http://www.brewerydb.com/styles}
}
\usage{
styles
}
\description{
BJCP style guidelines data from BreweryDB styles page
}
\keyword{datasets}
