% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tip.results.R
\name{tip.results}
\alias{tip.results}
\title{tip.results}
\usage{
tip.results(beer, what)
}
\arguments{
\item{beer}{the \code{"beer"} object}

\item{what}{which results to get}
}
\description{
Get results per tip
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
