% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.diff.R
\name{vector.diff}
\alias{vector.diff}
\title{vector differences}
\usage{
vector.diff(
  vector,
  base,
  return = c("angle", "projection", "rejection"),
  scaled = TRUE
)
}
\arguments{
\item{vector}{the vector to compare to the base (also a vector)}

\item{base}{the base vector}

\item{return}{what to return. Default is \code{c("angle", "projection", "rejection")}.}

\item{scale}{whether to scale the base vector to be the unit (\code{TRUE}).}
}
\description{
Comparing two vectors
}
\details{
This is a modification from \code{\link[dispRity]{projections}} to align the vector with the base and output all the three elements (angle, projection and rejection) rather than just one.
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
