% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar.utilities.R
\name{get.covar}
\alias{get.covar}
\alias{axes.covar}
\alias{ellipse.covar}
\title{Utilities for a dispRity object with covariance matrices}
\usage{
get.covar(data, n)

axis.covar(data, n, level = 0.95, axis = 1)
}
\arguments{
\item{data}{a \code{dispRity} object with a \code{covar} element.}

\item{sample}{optional, one or more specific posterior sample IDs (is ignored if n is used) or a function to summarise all axes.}

\item{n}{optional, a random number of covariance matrices to sample (if left empty, all are used).}

\item{level}{which confidence interval level to use (default is \code{0.95}).}

\item{axis}{which major axis to calculate (default is \code{1}, the first one).}
}
\description{
Different utility functions to extract aspects of a \code{MCMCglmm} object.
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
