% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.covar.R
\name{get.covar}
\alias{get.covar}
\title{Get covariance matrices}
\usage{
get.covar(data, n = 1)
}
\arguments{
\item{data}{The MCMCglmm object}

\item{n}{The number of samples}
}
\value{
Returns a \code{"beer"} object that is a list of length \code{n}.
Each \code{n}th element is a list of of the same length as the number of random + residual terms in \code{data} containing one variance covariance matrix (\code{get.covar(data)[[1]][[1]]$VCV}) and the intercept ((\code{get.covar(data)[[1]][[1]]$Sol})).
}
\description{
Get a sub-sample of the covariance matrices from a MCMCglmm object and the associated MME posterior intercept (solution) sorted by levels
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme, Gavin Thomas
}
