% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.covar.R
\name{get.covar}
\alias{get.covar}
\title{get.covar}
\usage{
get.covar(MCMCglmm, n = 1)
}
\arguments{
\item{MCMCglmm}{a MCMCglmm object}

\item{n}{how many covariance matrices to sample (default is \code{1})}
}
\value{
A \code{"beer"} object
}
\description{
sample covariance matrices from a MCMCglmm object
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
