% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.axes.R
\name{plot.axes}
\alias{plot.axes}
\title{plot.axes}
\usage{
\method{plot}{axes}(beer, axes, ..., add)
}
\arguments{
\item{beer}{the \code{"beer"} object.}

\item{axes}{which major axes to plot (default is \code{1}).}

\item{...}{any plotting options to be passed to \code{\link{graphics}{lines}}}

\item{add}{logical, whether to add the lines to an existing plot (\code{TRUE}, default) or not (\code{FALSE}).}
}
\description{
Plotting axes from an ellipse
}
\examples{

}
\seealso{

}
\author{
Thomas Guillerme
}
